# internal/controllers/swaggerdocs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для бронирования переговорок",
        "title": "Booking API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/rooms/{roomId}/schedule/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Create room schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "schedule payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.createScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.scheduleCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    }
                }
            }
        },
        "/rooms/{roomId}/slots/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "List available room slots by date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.slotsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.errorPayload"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.createScheduleRequest": {
            "type": "object",
            "properties": {
                "daysOfWeek": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "endTime": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
...
        "controllers.errorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controllers.errorPayload": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/controllers.errorBody"
                }
            }
        },
...
        "controllers.scheduleCreateResponse": {
            "type": "object",
            "properties": {
                "schedule": {
                    "$ref": "#/definitions/controllers.scheduleResponse"
                }
            }
        },
        "controllers.scheduleResponse": {
            "type": "object",
            "properties": {
                "daysOfWeek": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "controllers.slotResponse": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "controllers.slotsListResponse": {
            "type": "object",
            "properties": {
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.slotResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}